Use BitVectorIfc

+++ src/Rsdic.cpp
@@ -40,7 +40,7 @@ void Rsdic::clear()
     _one_num = 0;
 }
 
-bool Rsdic::get_bit(size_t pos) const
+bool Rsdic::get_bit(uint64_t pos) const
 {
     uint64_t lblock = pos / kLargeBlockSize;
     uint64_t pointer = _pointer_blocks[lblock];
@@ -53,7 +53,7 @@ bool Rsdic::get_bit(size_t pos) const
     return EnumCoder::get_bit(code, rank_sb, pos % kSmallBlockSize);
 }
 
-uint64_t Rsdic::rank0(const size_t pos) const
+uint64_t Rsdic::rank0(const uint64_t pos) const
 {
     uint64_t lblock = pos / kLargeBlockSize;
     uint64_t pointer = _pointer_blocks[lblock];
@@ -73,7 +73,7 @@ uint64_t Rsdic::rank0(const size_t pos) const
     return pos - rank;
 }
 
-uint64_t Rsdic::rank1(const size_t pos) const
+uint64_t Rsdic::rank1(const uint64_t pos) const
 {
     uint64_t lblock = pos / kLargeBlockSize;
     uint64_t pointer = _pointer_blocks[lblock];
@@ -93,7 +93,7 @@ uint64_t Rsdic::rank1(const size_t pos) const
     return rank;
 }
 
-pair<uint64_t, uint64_t> Rsdic::get_bit_and_rank(const size_t pos) const
+pair<uint64_t, uint64_t> Rsdic::get_bit_and_rank(const uint64_t pos) const
 {
     uint64_t lblock = pos / kLargeBlockSize;
     uint64_t pointer = _pointer_blocks[lblock];
@@ -112,7 +112,7 @@ pair<uint64_t, uint64_t> Rsdic::get_bit_and_rank(const size_t pos) const
 }
 
 
-uint64_t Rsdic::select1(const size_t ind) const
+uint64_t Rsdic::select1(const uint64_t ind) const
 {
     uint64_t select_ind = ind / kSelectBlockSize;
     uint64_t lblock = _select_one_inds[select_ind];
@@ -135,7 +135,7 @@ uint64_t Rsdic::select1(const size_t ind) const
     return sblock * kSmallBlockSize + EnumCoder::select1(code, rank_sb, remain);
 }
 
-uint64_t Rsdic::select0(const size_t ind) const
+uint64_t Rsdic::select0(const uint64_t ind) const
 {
     uint64_t select_ind = ind / kSelectBlockSize;
     uint64_t lblock = _select_zero_inds[select_ind];
@@ -162,14 +162,14 @@ uint64_t Rsdic::select0(const size_t ind) const
 namespace {
 template <typename T>
 void _save(std::ostream& os, const std::vector<T>& vs) {
-  size_t size = vs.size();
+  uint64_t size = vs.size();
   os.write((const char*)&size, sizeof(size));
   os.write((const char*)&vs[0], sizeof(vs[0]) * size);
 }
 
 template <typename T>
 void _load(std::istream& is, std::vector<T>& vs) {
-  size_t size = 0;
+  uint64_t size = 0;
   is.read((char*)&size, sizeof(size));
   vs.resize(size);
   is.read((char*)&vs[0], sizeof(vs[0]) * size);
@@ -204,7 +204,8 @@ void Rsdic::load(std::istream& is)
 }
 
 // Used for fast loading
-void Rsdic::load(const void *buf, const size_t len)
+// TODO: Make it cleaner
+void Rsdic::load(const void *buf, const uint64_t len)
 {
     const uint8_t *ptr = reinterpret_cast<const uint8_t*>(buf);
 
@@ -214,7 +215,7 @@ void Rsdic::load(const void *buf, const size_t len)
     memcpy(&_one_num, ptr, sizeof(_one_num));
     ptr += sizeof(_one_num);
 
-    size_t size;
+    uint64_t size;
 
     memcpy(&size, ptr, sizeof(size));
     ptr += sizeof(size);
